create Category Service, edit some fields

+++ GetandTake/Models/DTOs/ListDTO/ProductsDTO.cs
@@ -6,10 +6,10 @@
         public int ProductID { get; set; }
         public string ProductName { get; set; }
         public string? QuantityPerUnit { get; set; }
-        public int? UnitPrice { get; set; }
-        public int? UnitsInStock { get; set; }
-        public int? UnitsOnOrder { get; set; }
-        public int? ReorderLevel { get; set; }
+        public decimal? UnitPrice { get; set; }
+        public short? UnitsInStock { get; set; }
+        public short? UnitsOnOrder { get; set; }
+        public short? ReorderLevel { get; set; }
         public bool? Discontinued { get; set; }
 
         public string? Category { get; set; }
